Validate that environment variables are set

+++ install/install.sh
@@ -1,16 +1,21 @@
 #!/bin/bash
 
-./vendor/leoloso/wp-install/install/set-wp-config.sh
+# Validate that all required environment variables are set
+./vendor/leoloso/wp-install/install/validate-env-variables.sh
+if [ $? -eq 0 ]
+then
+    ./vendor/leoloso/wp-install/install/set-wp-config.sh
 
-# Save state if WP is already installed or not on .temp file, then maybe install
-( wp core is-installed && echo $? > .temp ) || echo $? > .temp
-./vendor/leoloso/wp-install/install/maybe-install-wp.sh
+    # Save state if WP is already installed or not on .temp file, then maybe install
+    ( wp core is-installed && echo $? > .temp ) || echo $? > .temp
+    ./vendor/leoloso/wp-install/install/maybe-install-wp.sh
 
-# Maybe show success message
-IS_WP_ALREADY_INSTALLED=$(cat .temp) && ./install/post-maybe-install-wp.sh $IS_WP_ALREADY_INSTALLED
+    # Maybe show success message
+    IS_WP_ALREADY_INSTALLED=$(cat .temp) && ./install/post-maybe-install-wp.sh $IS_WP_ALREADY_INSTALLED
 
-# Maybe show footer
-./install/maybe-print-footer.sh $IS_WP_ALREADY_INSTALLED
+    # Maybe show footer
+    ./install/maybe-print-footer.sh $IS_WP_ALREADY_INSTALLED
 
-# Delete the .temp file
-rm .temp
+    # Delete the .temp file
+    rm .temp
+fi
